requirements.txt
openpyxl==3.0.5
pandas==1.1.3
Pillow==9.3.0
plotly==5.11.0
streamlit==1.14.1
